+++ QuickFIXn/Logger/NonSessionLog.cs
@@ -1,9 +1,12 @@
+using System;
+
 namespace QuickFix.Logger;
 
 /// <summary>
 /// A logger that can be used when the calling logic cannot identify a session (which is rare).
 /// Does not create a log artifact until first write.
 /// </summary>
+[Obsolete("No longer needed.  Will be deleted in v1.16.")]
 public class NonSessionLog {
 
     private readonly ILogFactory _logFactory;
@@ -32,4 +35,3 @@ public class NonSessionLog {
         }
     }
 }
-