+++ QuickFIXn/Logger/ScreenLog.cs
@@ -1,4 +1,6 @@
-﻿namespace QuickFix.Logger;
+﻿using System;
+
+namespace QuickFix.Logger;
 
 /// <summary>
 /// FIXME - needs to log sessionIDs, timestamps, etc.
@@ -26,7 +28,7 @@ public class ScreenLog : ILog
         if (!_logIncoming)
             return;
 
-        System.Console.WriteLine("<incoming> " + msg.Replace(Message.SOH, '|'));
+        Console.WriteLine("<incoming> " + msg.Replace(Message.SOH, '|'));
     }
 
     public void OnOutgoing(string msg)
@@ -34,7 +36,7 @@ public class ScreenLog : ILog
         if (!_logOutgoing)
             return;
 
-        System.Console.WriteLine("<outgoing> " + msg.Replace(Message.SOH, '|'));
+        Console.WriteLine("<outgoing> " + msg.Replace(Message.SOH, '|'));
     }
 
     public void OnEvent(string s)
@@ -42,7 +44,7 @@ public class ScreenLog : ILog
         if (!_logEvent)
             return;
 
-        System.Console.WriteLine("<event> " + s);
+        Console.WriteLine("<event> " + s);
     }
     #endregion
 
@@ -54,7 +56,8 @@ public class ScreenLog : ILog
     }
     protected virtual void Dispose(bool disposing)
     {
-        // Nothing to dispose of...
+        // Nothing to dispose of
     }
+    ~ScreenLog() => Dispose(false);
     #endregion
 }