+++ QuickFIXn/ThreadedSocketAcceptor.cs
@@ -1,10 +1,11 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Net;
-using System;
+using System.Threading;
+using Microsoft.Extensions.Logging;
 using QuickFix.Logger;
 using QuickFix.Store;
-using System.Threading;
 
 namespace QuickFix
 {
@@ -22,15 +23,16 @@ namespace QuickFix
         private readonly SessionFactory _sessionFactory;
         private bool _disposed = false;
         private readonly object _sync = new();
-        private readonly NonSessionLog _nonSessionLog;
+        private readonly IQuickFixLoggerFactory _qfLoggerFactory;
+        private readonly LogFactoryAdapter? _logFactoryAdapter;
 
         /// <summary>
-        /// Create a ThreadedSocketAcceptor
+        /// Create a ThreadedSocketAcceptor (with a legacy ILogFactory)
         /// </summary>
         /// <param name="application"></param>
         /// <param name="storeFactory"></param>
         /// <param name="settings"></param>
-        /// <param name="logFactory">If null, a NullFactory will be used.</param>
+        /// <param name="logFactory">If null, a NullQuickFixLoggerFactory (which produces no logs) will be used.</param>
         /// <param name="messageFactory">If null, a DefaultMessageFactory will be created (using settings parameters)</param>
         public ThreadedSocketAcceptor(
             IApplication application,
@@ -38,12 +40,57 @@ namespace QuickFix
             SessionSettings settings,
             ILogFactory? logFactory = null,
             IMessageFactory? messageFactory = null)
+            : this(
+                application,
+                storeFactory,
+                settings,
+                logFactory is null ? NullQuickFixLoggerFactory.Instance : new LogFactoryAdapter(logFactory),
+                messageFactory)
+        { }
+
+        /// <summary>
+        /// Create a ThreadedSocketAcceptor
+        /// </summary>
+        /// <param name="application"></param>
+        /// <param name="storeFactory"></param>
+        /// <param name="settings"></param>
+        /// <param name="loggerFactory">If null, a NullQuickFixLoggerFactory (which produces no logs) will be used.</param>
+        /// <param name="messageFactory">If null, a DefaultMessageFactory will be created (using settings parameters)</param>
+        public ThreadedSocketAcceptor(
+            IApplication application,
+            IMessageStoreFactory storeFactory,
+            SessionSettings settings,
+            ILoggerFactory? loggerFactory = null,
+            IMessageFactory? messageFactory = null)
+            : this(
+                application,
+                storeFactory,
+                settings,
+                loggerFactory is null
+                    ? NullQuickFixLoggerFactory.Instance
+                    : new MelQuickFixLoggerFactory(loggerFactory),
+                messageFactory)
+        { }
+
+        private ThreadedSocketAcceptor(
+            IApplication application,
+            IMessageStoreFactory storeFactory,
+            SessionSettings settings,
+            IQuickFixLoggerFactory qfLoggerFactory,
+            IMessageFactory? messageFactory = null)
         {
-            ILogFactory lf = logFactory ?? new NullLogFactory();
+            if (qfLoggerFactory is LogFactoryAdapter lfa)
+            {
+                // LogFactoryAdapter is only ever created in the constructor marked obsolete, which means we own it and
+                // must save a ref to it so we can dispose it later. Any other loggerFactory is owned by someone else
+                // so we'll leave the dispose up to them. This should be removed eventually together with the old ILog
+                // and ILogFactory.
+                _logFactoryAdapter = lfa;
+            }
             IMessageFactory mf = messageFactory ?? new DefaultMessageFactory();
             _settings = settings;
-            _sessionFactory = new SessionFactory(application, storeFactory, lf, mf);
-            _nonSessionLog = new NonSessionLog(lf);
+            _sessionFactory = new SessionFactory(application, storeFactory, qfLoggerFactory, mf);
+            _qfLoggerFactory = qfLoggerFactory;
 
             try
             {
@@ -92,7 +139,7 @@ namespace QuickFix
 
             if (!_socketDescriptorForAddress.TryGetValue(socketEndPoint, out var descriptor))
             {
-                descriptor = new AcceptorSocketDescriptor(socketEndPoint, socketSettings, _nonSessionLog);
+                descriptor = new AcceptorSocketDescriptor(socketEndPoint, socketSettings, _qfLoggerFactory);
                 _socketDescriptorForAddress[socketEndPoint] = descriptor;
             }
 
@@ -174,7 +221,8 @@ namespace QuickFix
                 }
                 catch (Exception e)
                 {
-                    session.Log.OnEvent($"Error during logon of Session {session.SessionID}: {e.Message}");
+                    session.Log.Log(LogLevel.Error, e, "Error during logon of Session {SessionID}: {Message}",
+                        session.SessionID, e.Message);
                 }
             }
         }
@@ -189,7 +237,8 @@ namespace QuickFix
                 }
                 catch (Exception e)
                 {
-                    session.Log.OnEvent($"Error during logout of Session {session.SessionID}: {e.Message}");
+                    session.Log.Log(LogLevel.Error, e, "Error during logout of Session {SessionID}: {Message}",
+                        session.SessionID, e.Message);
                 }
             }
 
@@ -226,7 +275,8 @@ namespace QuickFix
                 }
                 catch (Exception e)
                 {
-                    session.Log.OnEvent($"Error during disconnect of Session {session.SessionID}: {e.Message}");
+                    session.Log.Log(LogLevel.Error, e, "Error during disconnect of Session {SessionID}, {Message}",
+                        session.SessionID, e.Message);
                 }
             }
         }
@@ -393,6 +443,7 @@ namespace QuickFix
             {
                 Stop();
                 DisposeSessions();
+                _logFactoryAdapter?.Dispose();
                 _sessions.Clear();
                 _disposed = true;
             }